Changes the code to adhere to the model
On the earlier code there were 2 variables added to PianoRoll that were related to the Window. That might be seem as bad, because it doesn't rigorously adhere to the model. That version solves that, with the cost of added complexity to the code. Submitted to an alternative branch for further review on which would be better for the PR.