Initial commit
	This commit makes a raw implementation of the logic that chooses
between the Automation Pattern values and the Controller values. The
behavior intended is that the Controller values will be used when the
song is stopped or when it's playing but there's an Automation being
recorded. The Automation Pattern values will be used otherwise.

	Things to improve:
	- isRecording() could be substituted by a member variable that
is set through Signals emitted from the AutomationPattern when it starts
 or stops recording for better performance.
	- isRecording() doesn't account for the position of the
Automation Pattern.
	- Even when there's no Automation Pattern linked to the model,
useControllerValues() return false when the song is playing.
AutomatableModel::isAutomated() could fix it, but calling it constantly
might not be the best way in terms of performance.