Improves performance when checking for recording
	Instead of looping through all automation patterns that are
connected to the automatable model and checking if they are recording,
we now have a member variable m_isRecording and a method called
updateIsRecording(). updateIsRecording() will loop through all
automation patterns and check if any is recording, but it will only be
called when any of the automation patterns connected to the model call
setRecording(). That way we reduce the overhead of cycling through the
patterns every time we call useControllerValues().