Improves performance on isAutomated()
	AutomatableModel::isAutomated() used to call
AutomationPattern::isAutomated() everytime it was called, which resulted
in it looping through all automation patterns that are connected to the
model. For real time usage, this might not be the best approach in terms
of performance.
	Instead, we are now using a m_isAutomated variable, which is
updated only when necessary (when AutomationPattern::addObject or
AutomationPattern::removeObject are called). The m_isAutomated variable
is then updated by updateIsAutomated() and we only retrieve its value
through isAutomated().
	For consistency, we also add a isRecording() method as a getter
for m_isRecording, which behaves similarly.

	This reduces the amount of times we will loop through the
AutomationPatterns by doing it only when we connect or disconnect a
model.

	m_isRecording and m_isAutomated are made mutable so we can call
their update methods from const objects.