Fixes isRecording and isAutomated
	There were a few situations where isRecording and isAutomated
weren't being updated when they should:
	- isRecording wasn't being updated when the model was connected
or disconnected to another automation pattern. Also wasn't being updated
when an automation pattern was destroyed.
	- isAutomated wasn't being updated when an automation pattern
was destroyed.

	To fix this, we added a destructor to the AutomationPattern
class that disconnects every model and calls their update methods. We
also added the updateIsRecording call to addObject and removeObject.

	Finally, we now check if the model is automated at all when
deciding between the controller values or the automation values.